fix: replace missmatched data types

+++ lib/Buffer/Buffer.cpp
@@ -1,7 +1,7 @@
 #include "Buffer.h"
 #include <stdio.h>
 
-void addToBuffer(int16_t value, struct CycleBuffer *buff) {
+void addToBuffer(int64_t value, struct CycleBuffer *buff) {
   buff->buffer[buff->head] = value;
   buff->head = (buff->head + 1) % BUFFER_SIZE;
   if (buff->count < BUFFER_SIZE) {
@@ -12,7 +12,7 @@ void addToBuffer(int16_t value, struct CycleBuffer *buff) {
 void logBuff(struct CycleBuffer *buff) {
   printf("HEAD: %d, COUNT: %d, BUFF: ", buff->head, buff->count);
   for(int i = 0; i < BUFFER_SIZE; i++) {
-    printf("%.2f ", buff->buffer[i]);
+    printf("%d ", buff->buffer[i]);
   }
   printf("\n");
 }
